Added finalization code.
Also added a try-catch-finally nest. Now, errors should be caught and the program itself will finalize instead of not completing at all. This is important because it ensures that the flow of water going out to the plants will not be left going in the event that an error occurs.